Add files via upload
A program that simulates a group of 6 students from several parallel classes and their academic results in four subjects. The program should read data for six students with the following properties:

- student's name;
- grades in computer science, mathematics, biology, and Polish language

Then it answers questions about the grade in a specific subject, where it first takes the student's number from the user and then the subject number.